Sync my project with the base project for starting the Advanced Android Programming course.
I decided it best to start from their code because I've made so many little deviations in the first class that instructions in the second class would likely conflict very much with my code. I'm sure I'll deviate in this class too but at least I'll be fresh on the changes I made so I don't be as confused.